% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corGaussian.R, R/corGaussianPeriodic.R
\name{corGaussian}
\alias{corCombined}
\alias{corGaussian}
\alias{corGaussianPeriodic}
\alias{corGaussianPeriodic2}
\alias{corMatern2.5}
\title{Correlation functions}
\usage{
corGaussian(inputs, inputs2, phi)

corMatern2.5(inputs, inputs2, phi)

corGaussianPeriodic(inputs, inputs2, phi, period)

corGaussianPeriodic2(inputs, inputs2, phi, period)

corCombined(inputs, inputs2, phi, cor.funcs, ...)
}
\arguments{
\item{inputs}{A data frame, matrix or vector containing the input values of the data where each row gives the observations.}

\item{inputs2}{An additional data frame, matrix or vector containing the input values of the data where each row gives the observations.}

\item{phi}{Estimate of the roughness parameter (a vector).}

\item{period}{A scalar or a vector indicating the period for each input parameter. If scalar, period for all input 
parameters will be set to the same scalar value as specified. Set period such that \eqn{cor(x_i, x_i + period) = 1}}

\item{cor.funcs}{An ordered vector of length equal to number of inputs containing characters \code{'g'}, \code{'m2.5'}, \code{'gp'}, or \code{'gp2'}
as an indicator for whether a particular input should be calculated using 
\code{corGaussian}, \code{corMatern2.5}, \code{corGaussianPeriodic}, or \code{corGaussianPeriodic2} respectively.}

\item{...}{Additional arguments to be passed on to correlation functions.}
}
\value{
Returns the correlation matrix \eqn{A} calculated by the specific function.
}
\description{
Functions to calculate the correlation between two points. 
The correlation functions give the correlation between the simulator output at any given vector of input values and 
the output at any other given input vector. Within this package, when these two vectors are the same then the correlation
function \code{C} at input vector \code{x} is \code{C(x,x) = 1}
In principal, the correlation function can take a very wide variety of forms. Only a few are included in this package, however, 
a user is able to define his own function. A user defined function must follow the same signature as those included in this package.
}
\details{
\code{corGaussian} uses the Gaussian correlation function defined as 
\deqn{\prod _{i=1} ^ {n} \exp [- (\frac{(x_i - x_i')}{\delta _i}) ^2]}{\prod _{i=1}^{n} exp(-((x_i - x_i') / \delta_i)^2)}

\code{corMatern2.5} uses the Matern correlation function with \eqn{\nu = 2.5} defined as 
\deqn{(1 + 5^{0.5}r + \frac{5}{3}r^2)\exp(-5^{0.5}r)}{(1 + (5^0.5)*r + (5*r^2)/3) * exp(-(5^0.5)*r)} where \eqn{r} is distance between inputs \eqn{x_i} and \eqn{x_j}, scaled by
delta: \deqn{[\frac{(x_{i1}-x_{j1})^2}{\delta _1^2} +...+ \frac{(x_{id}-x_{jd})^2}{\delta_d^2}]^{0.5}}{[(x_{i1}-x_{j1})^2 / \delta_1^2 +...+ (x_{id}-x_{jd})^2 / \delta_d^2 ]^{0.5}}

\code{corGaussianPeriodic} uses the Gaussian periodic correlation function defined as 
\deqn{\exp(- \delta ^2 sin(\frac{\pi}{period} |x_i - x_j|)^2)}{exp( - \delta ^2 * sin (\pi/period * |(x_i - x_j)|)^2)}
\code{corGaussianPeriodic2} uses a different way of accounting for 
the periodicity in the data. Both \code{corGaussianPeriodic} and \code{corGaussianPeriodic2} give the same results.

\code{corCombined} is a function that allows the user to specify a different correlation function for each output 
(from the list of predefined functions).
}
\note{
These functions use \code{\link{rdist}} from package \pkg{fields} for speed.
}
\author{
Sajni Malde and Jeremy Oakley
}
\seealso{
See \code{\link{optim}} or  \code{\link{MCMCMetropolisGibbs}} where the function is used.
}

