% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal.R
\name{crossVal}
\alias{crossVal}
\title{Cross Validation}
\usage{
crossVal(fit, lmcompare = FALSE, lmformula = fit$formula, plot = TRUE)
}
\arguments{
\item{fit}{Object of class \code{emulatorFit} generated by \code{\link{fitEmulator}} or \code{\link{fitEmulatorData}}}

\item{lmcompare}{Compare with a linear model fitted with OLS.}

\item{lmformula}{The formula for the linear model to be used in the comparison, if \code{lmcompare} is \code{TRUE}.}
}
\value{
This function returns the mean RMSE of fits where one training point is left out for each fit.
}
\description{
Leave-one-out cross validation
}
\note{
This function requires R package \pkg{Hmisc} to be installed.
}
\author{
Jeremy Oakley.
}
