% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emPlot1D.R
\name{emPlot1D}
\alias{emPlot1D}
\title{emPlot1D}
\usage{
emPlot1D(fit, input = 1, lower = NULL, upper = NULL, alpha = 0.05)
}
\arguments{
\item{fit}{Object of class \code{emulatorFit} generated by \code{\link{fitEmulator}} or \code{\link{fitEmulatorData}}}

\item{input}{The default is set to 1}

\item{lower}{The default is set to Null}

\item{upper}{The default is set to Null}

\item{alpha}{The default is set to 0.05}
}
\value{
Returns a plot of a single input against output
}
\description{
Plot emulator distribution across a single input
}
\note{
This function requires R package \pkg{ggplot2} to be installed.
}
\author{
Jeremy Oakley.
}

