% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEmulatorData.R
\name{fitEmulatorData}
\alias{fitEmulatorData}
\title{Fit Emulator (Data)}
\usage{
fitEmulatorData(formula, data, input.names = NULL, ...)
}
\arguments{
\item{formula}{Formula to include the RHS and LHS. The LHS indicates the outputs and the RHS defines the prior mean function defined by the column names of \code{data}. 
A multivariate formula can also be given (see \code{\link{as.Formula}}) .}

\item{data}{A data frame or matrix containing at least the inputs and outputs of the model.}

\item{input.names}{An optional vector of character containing the names of variables in data to be considered as input variables. 
By default, all variables in the data matrix, except those on the left hand side of the formula, are treated as inputs.}

\item{...}{Additional arguments to be passed to the more specific \code{fitEmulator} function.
Arguments such as \code{MCMC.iterations}, \code{phi.opt}, \code{MC.plot}, etc.}
}
\value{
\code{fitEmulatorData} returns an object of class 'emulatorFit'. 
         See \code{\link{fitEmulator}} for details.
}
\description{
A function equivalent to \code{fitEmulator}. In \code{fitEmulatorData} the user can provide
the inputs and outputs in one matrix/dataframe. To provide them seperately, use \code{\link{fitEmulator}}
}
\examples{
# specify the prior mean function as a object of class 'formula'
# ensure colnames of datasets include at least these variables.
emulator.prior.mean.formula <- Y ~ X.1 + X.2  

# fit the emulator using a prior mean specified above and a Gaussian correlation function.
fit <- fitEmulatorData(formula = emulator.prior.mean.formula, data = surfebm[1:25, ], 
                       cor.function = corGaussian)

# Use fitted emulator to predict posterior means and variances at the new inputs
predictions <- predict(fit, surfebm[26:35, 1:2], sd = FALSE, var.cov = TRUE)

# view results
print(predictions$posterior.mean)
}
\seealso{
If input and output data is in different objects, see \code{\link{fitEmulator}} as a more conveniant method.
}

