% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report for the GPE analysis}
\usage{
generateReport(fit, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to the \code{\link{render}} function (from rmarkdown).}

\item{emulator}{A fit object of class inheriting from \code{'emulatorFit'}.}

\item{new.inputs}{A data matrix of input(s) at which emulation is desired (new inputs). 
Must contain at least all parameters given in \code{emulator$training.inputs}.
(Ensure they are the same used to generate the \code{emulator.predictions}.)}

\item{new.outputs}{A matrix outputs used to validate the emulator predictions. This Matrix must have columns equal to the number of outputs.}

\item{emulator.predictions}{Output generated from the \code{\link{predict.emulatorFit}} function. It is an optional argument. 
If this is not provided it will be calculated using arguments provided. Note Posterior Variance will be calculated, unless you provide this argument.}
}
\description{
Generate a report for the GPE analysis
}
