% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSample.R
\name{GPSample}
\alias{GPSample}
\title{GP Sample}
\usage{
GPSample(n, mean, var)
}
\arguments{
\item{n}{the number of samples required}

\item{mean}{a vector giving the means of the variables}

\item{var}{a positive-definite symmetric matrix specifying the covariance matrix of the variables}
}
\description{
Function to sample from posterior Gaussian process
}
\author{
Jeremy Oakley.
}

