% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEmulatorConstants.R
\name{makeEmulatorConstants}
\alias{makeEmulatorConstants}
\alias{makeEmulatorConstantsLMC}
\title{Make Emulator Constants}
\usage{
makeEmulatorConstants(phi, inputs, H, outputs, cor.function, nugget = NULL,
  sigmasq.hat = NULL, ...)

makeEmulatorConstantsLMC(phi, sigma, inputs, H, outputs, cor.function, ...)
}
\arguments{
\item{phi}{Optimum Maximum likelihood estimate of the correlation length parameter}

\item{inputs}{A data frame, matrix or vector containing the input values of the training data.}

\item{H}{A matrix of prior mean regressors for the training data}

\item{outputs}{A data frame, matrix or vector containing the output values of the training data.In \code{makeEmulatorConstantLMC} the outputs should be stacked (either a vector or a matrix with 1 column).}

\item{cor.function}{Specifies a correlation function used as part of the prior information for the emulator. 
This package has options of: \code{\link{corGaussian}}, \code{\link{corMatern2.5}}, \code{\link{corGaussianPeriodic}},
 \code{\link{corGaussianPeriodic2}} and \code{\link{corCombined}}. One can also specify a user defined function.}

\item{nugget}{For noisy data, a vector giving the observation variance for each training data point.}

\item{sigmasq.hat}{Optimum Maximum likelihood estimate of the unknown scale parameter.}

\item{...}{additional arguments to be passed on to correlation functions (see \code{\link{corGaussian}})}

\item{sigma}{Optimum Maximum likelihood estimate of the between outputs variance covariance matrix.}
}
\value{
The function returns a list containting some components as the \code{emulatorFit} class. 
         See \code{\link{fitEmulator}} for details
}
\description{
Function for collating and calculating constants needed for fitting the emulator
}
\author{
Originally written by Jeremy Oakley. Modified by Sajni Malde
}
\references{
J. Oakley 1999. 'Bayesian uncertainty analysis for complex computer codes', PhD thesis, University of Sheffield.
}

