% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMetropolisGibbs.R
\name{MCMCMetropolisGibbs}
\alias{MCMCMetropolisGibbs}
\title{MCMC using Metroplis Hastings within Gibbs}
\usage{
MCMCMetropolisGibbs(inputs, outputs, fn, H, MCMC.iterations, starting.values,
  proposal.sd = 0.1, cor.function, MC.plot = TRUE, ...)
}
\arguments{
\item{inputs}{A data frame, matrix or vector containing the input values of the training data.}

\item{outputs}{A data frame, matrix or vector containing the output values of the training data.}

\item{fn}{A function used to maximise the negetive log likelihood}

\item{H}{A matrix of prior mean regressors from the training data}

\item{MCMC.iterations}{The number of iterations that MCMC should be run for}

\item{starting.values}{the starting values for which the MCMC can start running}

\item{proposal.sd}{is the standard deviation of the random walk proposal (default \code{0.1})}

\item{cor.function}{Specifies a correlation function used as part of the prior information for the emulator. 
This package has options of: \code{\link{corGaussian}}, \code{\link{corMatern2.5}}, \code{\link{corGaussianPeriodic}},
 \code{\link{corGaussianPeriodic2}} and \code{\link{corCombined}}. One can also specify a user defined function.}

\item{MC.plot}{If \code{TRUE}, produces a trace plot of the MCMC output of log likelihood against the number of iterations. (default=\code{TRUE})}

\item{...}{additional arguments to be passed on to correlation functions (see \code{\link{corGaussian}})}
}
\value{
The function returns a list containting the following components: 
\tabular{ll}{
\code{density.sample} \tab  The negetive log likelihood of the MCMC output at the starting value and at each iteration \cr
\code{theta.sample} \tab  A matrix of the theta sample at the starting value and at each iteration   \cr }
}
\description{
Generate a Markov Chain of the parameters in the correlation function Using Metropolis-Hastings within Gibbs
}
\note{
Note that this function first calculates the negetive log likelihood of the starting values so returns \code{MCMC.iterations} + 1 values.
}
\author{
Originally written by Jeremy Oakley. Modified by Sajni Malde
}
