% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negLogLikGaussianRcpp.R
\name{negLogLikGaussianRcpp}
\alias{negLogLikGaussianRcpp}
\title{Negative Log Likelihood Functions}
\usage{
negLogLikGaussianRcpp(theta, inputs, H, outputs, nugget = NULL, ...)
}
\arguments{
\item{theta}{Initial values for the parameters to be optimized over. (includes values for all the parameters that need to be estimated.)
For \code{negLogLik}, theta represents just phi, for \code{negLogLikNugget}, theta represents just phi and sigma combined.}

\item{inputs}{A data frame, matrix or vector containing the input values of the training data.}

\item{H}{A matrix of prior mean regressors for the training data.}

\item{outputs}{A data frame, matrix or vector containing the output values of the training data. In \code{negLogLikLMCOptim}, the outputs should be stacked (either a vector or a matrix with 1 column).}

\item{nugget}{For noisy data, a vector giving the observation variance for each training data point.}

\item{...}{additional arguments to be passed on to correlation functions (see \code{\link{corGaussian}})}

\item{cor.function}{Specifies a correlation function used as part of the prior information for the emulator}
}
\value{
The function returns the negetive log-likelihood of \code{theta}
}
\description{
These functions are minimised to find the optimal value of theta (or phi and sigma in LMC case). 
             \code{negLogLikNugget} is used in the univariate case for uncertain input parameters.
             \code{negLogLikLMC} is used when applying the LMC Emulator for Multivariate models
}
\seealso{
\code{\link{corGaussian}}
}
\author{
Sajni Malde, Jeremy Oakley
}
