% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.emulatorFit.R
\name{predict.emulatorFit}
\alias{predict.emulatorFit}
\title{Prediction using Emulators}
\usage{
\method{predict}{emulatorFit}(object, newdata, var.cov = FALSE, sd = TRUE,
  tol = -1e-11, ...)
}
\arguments{
\item{object}{A fit object of class inheriting from \code{'emulatorFit'}.}

\item{newdata}{A data matrix of input(s) at which emulation is desired (new inputs). 
Must contain at least all parameters given in \code{object$training.inputs}.
If missing, the fitted inputs \code{object$training.inputs} are used.}

\item{var.cov}{Optionally calculates posterior variance covariance matrix. Default is set to FALSE. For large numbers of training and prediction data, this is quite time consuming.}

\item{sd}{Optionally calculates only the posterior standard deviation. Default is set to \code{TRUE}.}

\item{tol}{The tolerance for capping negative small values of posterior standard deviation to zero.
The default is \eqn{-10^{-11}}{-10^-11}.}

\item{...}{Further arguments not used and an error is thrown if provided.}
}
\value{
The function returns a list containting the following components: 
\tabular{ll}{
    \code{posterior.mean}  \tab    Approximation of the outputs for the given inputs in \code{newdata}   \cr
    \code{posterior.variance}  \tab Variance covariance matrix around this approximation \cr
    \code{standard.deviation} \tab  Standard Deviation of the approximation. It equals the square-root
                                   of the diagonal of the \code{posterior.variance}
}
When the number of outputs to emulate is more than 1, \code{method = 'separable'}, and object is of class \code{"emulatorFit"} 
two extra values are returned from this function. These are 
\tabular{ll}{
    \code{correlation.Matrix}  \tab A spatial correlation matrix.  \cr
    \code{sigmahat}  \tab    A between outputs covariance matrix.  \cr
}
}
\description{
Predicts value and confidence interval at new inputs using Gaussian Process Emulation. 
This function should be preceded by the \code{\link{fitEmulator}} function.
}
\details{
Note that when using the LMC method, calculating the posterior variance is quite time-consuming.
}
\author{
Originally written by Jeremy Oakley. Modified by Sajni Malde.
}
\references{
Oakley, J. (1999). Bayesian uncertainty analysis for complex computer codes, Ph.D. thesis, University of Sheffield.
}

