% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surfebm}
\alias{surfebm}
\title{surfebm}
\format{data.frame of 35 observations for 3 variables.}
\usage{
surfebm
}
\description{
This is an example dataset used in package.
}
\author{
Jeremy Oakley
}
