% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFit.R
\name{updateFit}
\alias{updateFit}
\title{Update fit}
\usage{
updateFit(object)
}
\arguments{
\item{object}{An \code{"emulatorFit"} object to update.}
}
\value{
Returns the updated \code{"emulatorFit"} object.
}
\description{
Updates old \code{"emulatorFit"} objects to work with latest 
code to allow for improvements to be made within the code.
}
\note{
A warning is issued if the object is updated.
}

