% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateEmulator.R
\name{validateEmulator}
\alias{validateEmulator}
\title{Validate the Emulator}
\usage{
validateEmulator(emulator, new.outputs, emulator.predictions,
  which.output = 1, plot = TRUE)
}
\arguments{
\item{emulator}{Object of class \code{emulatorFit} generated by \code{\link{fitEmulator}} or \code{\link{fitEmulatorData}}}

\item{new.outputs}{A matrix outputs used to validate the emulator predictions. This Matrix must have columns equal to the number of outputs.}

\item{emulator.predictions}{Output generated from the \code{\link{predict.emulatorFit}} function.}

\item{which.output}{Defaults to 1. Is an indicative integer that validates the output stated based on column number from the left.}

\item{plot}{Optionally plot validation plots. Default is set to true.}
}
\value{
The class of '\code{validateEmulator}' is a list containting atleast the following components: 
\tabular{ll}{
 \code{coverage} \tab Proportion of validation outputs in 95 percent confidence intervals of the emulator predicted output \cr 
 \code{RMSE}  \tab    The root mean square error of the predicted values   \cr
 \code{normRMSE} \tab The normalised value of \code{RMSE} over the range of the data \cr}
}
\description{
Uses diagnostics described in Bastos and O'Hagan (2009) to validate the emulator.
}
\details{
\code{validateEmulator} returns a grid of 4 plots and prints out a value. These are described below: 
\tabular{ll}{ 
Top Left plot \tab  Plots validation output against the posterior mean generated using the emulator. The error bars are the 95percent confidence intervals of the predicted mean. Ideally, the plot should show all points close to the diagonal line with small error bars. \cr
Top Right plot \tab Shows the QQ plot of cholesky residuals.   
If the points lie close to the 45 degree line then the normality assumptions for the simulator output is reasonable. If the gradient of the points is greater than 1 (less than 1), it suggests that the predictive variability was underestimated (over estimated). 
Curvature in the plot indicates nonnormality, and outliers at either end of the plot suggests local fitting problems or nonstationarity.    \cr
Bottom Left plot \tab Plot of Pivoted cholesky prediciton errors against the pivoting index. 
The pivoting index gives the order of the Pivoted cholesky prediciton errors with the largest conditional predictive variance. No patterns are expected. Too many large errors indicate an underestimation of variance and vice versa.Both cases can aslo suggest a nonstationary process. 
Either large or very small errors at the beginning of the plot (i.e., on the left side) indicates poor estimation of predictive variance or nonstationarity. 
However, large (or very small) errors at the end of the plot (i.e., on the right side) indicates overestimation (or underestimation) of the correlation length parameters, or that the chosen correlation structure is unsuitable. \cr
Bottom right Plot \tab The red line on the plot is the value of the calculated Mahalanobis distance taking into account correlation among outputs. Ideally it should be between the boundaries of the curve. Extreme values (large or small) indicate a conflict between the emulator and simulator.   \cr }
}
\author{
Originally written by Jeremy Oakley. Modified by Sajni Malde
}
\references{
Bastos, L. S. and O'Hagan, A. (2009). Diagnostics for gaussian process emulators, Technometrics, 51 (4): 425-438.
}
\seealso{
\code{\link{predict.emulatorFit}} and \code{\link{fitEmulator}}
}

