% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidataEmulatorApp.R
\name{validateEmulatorApp}
\alias{validateEmulatorApp}
\title{validating the emulator emulator}
\usage{
validateEmulatorApp(emulator, new.inputs, new.outputs,
  emulator.predictions = NULL, verbose = TRUE, ...)
}
\arguments{
\item{emulator}{A fit object of class inheriting from \code{'emulatorFit'}.}

\item{new.inputs}{A data matrix of input(s) at which emulation is desired (new inputs). 
Must contain at least all parameters given in \code{emulator$training.inputs}.
(Ensure they are the same used to generate the \code{emulator.predictions}.)}

\item{new.outputs}{A matrix outputs used to validate the emulator predictions. This Matrix must have columns equal to the number of outputs.}

\item{emulator.predictions}{Output generated from the \code{\link{predict.emulatorFit}} function. It is an optional argument. 
If this is not provided it will be calculated using arguments provided. Note Posterior Variance will be calculated, unless you provide this argument.}

\item{verbose}{Defaults to TRUE. If FALSE, text under graphs (explanations) will not appear.}
}
\description{
validating the emulator emulator
}
\examples{
# Fit the emulator
fit <- fitEmulator(inputs = surfebm[1:25, 1:2], 
                   outputs = surfebm[1:25, 3, drop = FALSE])

# Use fitted emulator to predict posterior means and variances at the new points
predictions <- predict(fit, surfebm[26:35, 1:2], sd = FALSE, var.cov = TRUE)

# Compare predictions with true values for the new inputs
# Can also compare accuracy of prediction based on posterior variance
validateEmulatorApp(fit, surfebm[26:35, 1:2], surfebm[26:35, 3], predictions, verbose = TRUE, launch.browser = TRUE)

}

